test(squad): promote VR/OA monoliths into tests/squad/<agent>/ packages
The big tests/test_squad_*.py monoliths predate the tests/squad/<agent>/ layout
that PT / PM / DC already use. Promote the two largest stragglers:

- VR (569 lines, one class) split along its two tool-module seam into
  tests/squad/vulnerability_researcher/test_tools.py (research-phase: NVD/CVE
  lookups, host service/release readers, per-OAM-asset vuln annotation,
  attack-forest finalisation) and test_triage.py (triage-phase: raw-finding
  readers, assess/discard, CVSS/CWE/OWASP lookups, triage finalisation).
- OA (349 lines) relocated whole to tests/squad/osint_analyst/test_tools.py
  (its enrichment tests already live in test_enrichment.py).

Pure relocation + split along an existing seam; no test bodies changed. Verified
in a CI-faithful py3.12 venv: ruff + mypy clean, unit suite green, diff-cover
100%.
