+++ tests/squad/vulnerability_researcher/test_triage.py
@@ -1,0 +1,212 @@
+"""
+tests/squad/vulnerability_researcher/test_triage.py - exercise the Vulnerability
+Researcher's triage-phase @tool wrappers (raw-finding readers, Assess / Discard
+Finding, CVSS / CWE / OWASP lookups, triage finalisation).
+
+The wrappers are thin: unmarshal JSON, call into tools/* helpers, serialise
+the result. Coverage here is regression coverage of the wrapping itself; the
+underlying helpers are exercised in their own dedicated test files. The
+research-phase wrappers live in ``test_tools.py``.
+"""
+
+from __future__ import annotations
+
+import json
+
+import pytest
+
+pytestmark = pytest.mark.unit
+
+
+class TestVulnerabilityResearcherTriageTools:
+    @staticmethod
+    def _write_findings(tmp_path, raw_finding_high) -> None:
+        (tmp_path / "findings.json").write_text(
+            json.dumps([raw_finding_high.model_dump(mode="json")]),
+            encoding="utf-8",
+        )
+
+    @staticmethod
+    def _good_authoring(target_apex: str, **overrides):
+        """Build the kwargs for ``assess_finding_tool.func(...)``.
+
+        The wrapper takes a typed ``AuthoredAssessment``, so the
+        authored fields are nested under ``authored`` while
+        ``finding_index`` / ``findings_path`` / ``programme_handle``
+        stay top-level. ``overrides`` mutate authored fields when the
+        key matches one, top-level otherwise.
+        """
+        authored = {
+            "severity_decision": "keep",
+            "severity": "high",
+            "severity_rationale": (
+                "Unauthenticated SQLi at a public endpoint with full DB read "
+                "available - matches the PT high call."
+            ),
+            "cvss_vector": "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:H/A:H",
+            "title": "SQL Injection in /search?q allows database extraction",
+            "description": (
+                "The /search endpoint concatenates q into a SELECT statement without "
+                "parameterisation. sqlmap exploits classic UNION-based injection to "
+                "extract rows from the users table."
+            ),
+            "steps_to_reproduce": [
+                f"GET https://api.{target_apex}/search?q=test' UNION SELECT 1,2,3-- ",
+                "Observe the response body contains the union'd rows.",
+            ],
+            "impact": (
+                "An authenticated attacker dumps the users table including bcrypt "
+                "hashes and emails, enabling offline cracking and account takeover."
+            ),
+            "remediation": (
+                "Use parameterised queries throughout. See "
+                "https://cheatsheetseries.owasp.org/cheatsheets/SQL_Injection_Prevention_Cheat_Sheet.html"
+            ),
+        }
+        base: dict[str, object] = {
+            "finding_index": 0,
+            "authored": authored,
+        }
+        for key, value in overrides.items():
+            if key in authored:
+                authored[key] = value
+            else:
+                base[key] = value
+        return base
+
+    def test_list_raw_findings_tool(self, raw_finding_high, run_dir) -> None:
+        from squad.vulnerability_researcher import list_raw_findings_tool
+
+        self._write_findings(run_dir, raw_finding_high)
+        result = list_raw_findings_tool.func()
+
+        assert isinstance(result, list)
+        assert result[0].index == 0
+        assert result[0].vuln_class == raw_finding_high.vuln_class
+        assert result[0].evidence_bytes == len(raw_finding_high.evidence)
+
+    def test_read_raw_finding_tool(self, raw_finding_high, run_dir) -> None:
+        from models import RawFinding
+        from squad.vulnerability_researcher import read_raw_finding_tool
+
+        self._write_findings(run_dir, raw_finding_high)
+        result = read_raw_finding_tool.func(0)
+
+        assert isinstance(result, RawFinding)
+        assert result.target == raw_finding_high.target
+
+    def test_read_raw_finding_tool_rejects_out_of_range(self, raw_finding_high, run_dir) -> None:
+        from squad.vulnerability_researcher import read_raw_finding_tool
+
+        self._write_findings(run_dir, raw_finding_high)
+        with pytest.raises(ValueError, match="out of range"):
+            read_raw_finding_tool.func(5)
+
+    def test_lookup_cwe_tool_finds_known_class(self) -> None:
+        from models import CWE
+        from squad.vulnerability_researcher import lookup_cwe_tool
+
+        result = lookup_cwe_tool.func(89)
+        assert isinstance(result, list)
+        assert result
+        assert isinstance(result[0], CWE)
+        assert result[0].cwe_id == 89
+
+    def test_lookup_owasp_tool_returns_cheatsheet(self) -> None:
+        from squad.vulnerability_researcher import lookup_owasp_tool
+        from tools.owasp_data import OWASPEntry
+
+        result = lookup_owasp_tool.func("sql injection")
+        assert isinstance(result, list)
+        assert all(isinstance(r, OWASPEntry) for r in result)
+        assert any("SQL_Injection_Prevention" in r.url for r in result)
+
+    def test_calculate_cvss_tool(self) -> None:
+        from squad.vulnerability_researcher import calculate_cvss_tool
+
+        result = calculate_cvss_tool.func("CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H")
+        assert result == 9.8
+
+    def test_assess_finding_tool_writes_assessment_and_returns_validation(
+        self, raw_finding_high, programme_in_workspace, run_dir, target_apex
+    ) -> None:
+        from squad.vulnerability_researcher import assess_finding_tool
+
+        self._write_findings(run_dir, raw_finding_high)
+        result = assess_finding_tool.func(**self._good_authoring(target_apex))
+
+        from tools.triage_tools import AssessmentResult
+
+        assert isinstance(result, AssessmentResult)
+        assert result.validation.ok is True
+        assert (run_dir / "assessments" / "000.json").exists()
+
+    def test_assess_finding_tool_surfaces_validation_issues(
+        self, raw_finding_high, programme_in_workspace, run_dir, target_apex
+    ) -> None:
+        from squad.vulnerability_researcher import assess_finding_tool
+
+        self._write_findings(run_dir, raw_finding_high)
+        result = assess_finding_tool.func(
+            **self._good_authoring(target_apex, description="Too short.")
+        )
+
+        from tools.triage_tools import AssessmentResult
+
+        assert isinstance(result, AssessmentResult)
+        assert result.validation.ok is False
+        sections = {i.section for i in result.validation.issues}
+        assert "description" in sections
+
+    def test_discard_finding_tool_writes_discard(self, raw_finding_high, run_dir) -> None:
+        from squad.vulnerability_researcher import discard_finding_tool
+
+        self._write_findings(run_dir, raw_finding_high)
+        result = discard_finding_tool.func(
+            finding_index=0,
+            reason="false_positive",
+            rationale="On reproduction the tool's marker did not appear.",
+        )
+
+        from tools.triage_tools import DiscardResult
+
+        assert isinstance(result, DiscardResult)
+        assert result.path == "discards/000.json"
+        assert (run_dir / "discards" / "000.json").exists()
+
+    def test_discard_finding_tool_rejects_thin_rationale(self, raw_finding_high, run_dir) -> None:
+        from squad.vulnerability_researcher import discard_finding_tool
+
+        self._write_findings(run_dir, raw_finding_high)
+        with pytest.raises(ValueError, match="at least 10 chars"):
+            discard_finding_tool.func(finding_index=0, reason="false_positive", rationale="nope")
+
+    def test_finalise_triage_tool_consolidates_assessments(
+        self, raw_finding_high, programme_in_workspace, run_dir, target_apex
+    ) -> None:
+        from squad.vulnerability_researcher import assess_finding_tool, finalise_triage_tool
+
+        self._write_findings(run_dir, raw_finding_high)
+        assess_finding_tool.func(**self._good_authoring(target_apex))
+        result = finalise_triage_tool.func()
+
+        assert result == "verified.json"
+        assert (run_dir / "verified.json").exists()
+
+    def test_finalise_triage_tool_raises_on_unprocessed(
+        self, raw_finding_high, programme_in_workspace, run_dir
+    ) -> None:
+        from squad.vulnerability_researcher import finalise_triage_tool
+
+        # write TWO raw findings but assess none
+        (run_dir / "findings.json").write_text(
+            json.dumps(
+                [
+                    raw_finding_high.model_dump(mode="json"),
+                    raw_finding_high.model_dump(mode="json"),
+                ]
+            ),
+            encoding="utf-8",
+        )
+        with pytest.raises(ValueError, match="no assessment or discard"):
+            finalise_triage_tool.func()
